+++ CHANGES.md
@@ -283,6 +283,11 @@ OpenSSL Releases
 
    *Dimitri John Ledkov*
 
+ * Added optimized ML-KEM NTT and scalar multiplication operations on `s390x`
+   (with potential reuse for other architectures with 128 bit vector registers).
+
+   *Timo Keller*
+
 OpenSSL 4.0
 -----------